Delete the preview file no matter what happened (bug 1051854)

+++ apps/devhub/tasks.py
@@ -233,11 +233,13 @@ def resize_preview(src, instance, **kw):
                                       remove_src=False)
         instance.sizes = sizes
         instance.save()
-        # Finally delete the temporary now useless source file.
-        os.unlink(src)
         return True
     except Exception, e:
         log.error("Error saving preview: %s" % e)
+    finally:
+        # Finally delete the temporary now useless source file.
+        if os.path.exists(src):
+            os.unlink(src)
 
 
 @task
